python/requirements.txt
httpx[http2]
Pillow
numpy
joblib
